02-memory-rag-agent/requirements.txt
# Standalone dependencies for 02 - Memory + RAG Agent
openai
python-dotenv
pypdf
